+++ example/btc_sdk_example.dart
@@ -1,65 +1,39 @@
-import 'dart:convert';
-import 'dart:ffi';
-import 'dart:typed_data';
-
 import 'package:btc_sdk/btc_sdk.dart';
-import 'package:fast_base58/fast_base58.dart';
-import 'package:test/test.dart';
 
 void main() {
-  group('btc_sdk tests', ()
-  {
 
-    final String myMnemonic = 'card file race stamp craft behave pulp achieve security grace leopard recall';
+  // Provided the mnemonic, creates the associated HDWallet
+  final String mnemonic = 'card file race stamp craft behave pulp achieve security grace leopard recall';
+  final HDWallet wallet = HDWallet.fromMnemonic(mnemonic);
 
-    setUp(() {
-      // Additional setup goes here.
-    });
+  final wif = wallet.masterPrivateKey.toWifPrivateKey(Network.mainnet, true); // L3tPgshYhzxDFf9EyGdCZEyZcQ6LqtxV5CmRy5q178EokrhRQqyU
 
-    test('Uint8List data types', () {
-      final Uint8List data = Uint8List.fromList([124, 8, 234, 156]);
-      final int first = data[0];
-      expect(first, 0x7C, reason: 'First byte is 0x7C = 124');
-      int value4Bytes = data.buffer.asByteData().getInt32(0); // read the full Uint8List of length 4 as a Uint32 (4 bytes int value)
-      expect(value4Bytes, 2080959132, reason: 'Read the full Uint8List of length 4 as an int32 value: 2080959132');
-      expect(2080959132, 0x7C08EA9C, reason: "2080959132 can be represented as the concat of the [124 = 7C, 8 = 08, 234 = EA, 156 = 9C] in HEX 7C08EA9C");
+  print("Wallet Inport Format for the HD Wallet: ${wif.toWif}");
 
-      final String hexData = data.toHex;
-      expect(hexData, '7C08EA9C'.toLowerCase());
-      expect('7C08EA9C'.toUint8ListFromHex, [124, 8, 234, 156]);
+  final ExtendedPublicKey pubKey = wallet.masterPrivateKey.extendedPublicKey;
 
-      expect(hexData.toUint8ListFromHex, data);
-      expect('hello'.toUint8ListFromHex, null, reason: 'An invalid HEX string should generate a null Uint8List');
+  final p2pkhAddr = pubKey.toAddress().p2pkhBase58;
 
-      expect(data.toBase58, '4AtTej');
-      expect('3yQ'.toUint8ListFromBase58?.buffer.asByteData().getInt16(0), 9999);
+  print("Pay to Public Key Hash address: ${p2pkhAddr}"); // 1HnX4BnhhYXQNMBGPMVDABE4HHYjanDRCh
+  print("Address length: ${p2pkhAddr.length}"); // length = 34
 
-      final l1 = [128].toUint8List;
-      final l2 = [64].toUint8List;
-      final l3 = l1.concat(l2);
-      expect(l3, [128, 64]);
+  final bitcoinTransactionBytes = "020000000255a736179f5ee498660f33ca6f4ce017ed8ad4bd286c162400d215f3c5a876af000000001976a9140af2c7db949861cdc7767ad211432789f1852e9488acffffffff4d89764cf5490ac5023cb55cd2a0ecbfd238a216de62f4fd49154253f1a750920200000000ffffffff0200e20400000000001976a914e993470936b573678dc3b997e56db2f9983cb0b488ac20cb0000000000001976a914b780d54c6b03b053916333b50a213d566bbedd1388ac0000000001000000".toUint8ListFromHex!;
 
-      final l4 = l3.appendInt(33);
-      expect(l4, [128, 64, 33]);
+  final Transaction transaction = TransactionReader.fromUint8List(bitcoinTransactionBytes).readTransaction();
 
-      final l0 = Uint8List(0);
-      final l16 = l0.appendInt(Uint.minUint16Value);
-      expect(l16, [1, 0]);
-      expect(l16?.toHex, '0100');
-      final l32 = l0.appendInt(Uint.minUint32Value);
-      expect(l32, [0, 1, 0, 0]);
-      expect(l32?.toHex, '00010000');
-      final l64 = l0.appendInt(Uint.minUint64Value);
-      expect(l64, [0, 0, 0, 1, 0, 0, 0, 0]);
-      expect(l64?.toHex, '0000000100000000');
-    });
+  print("Transaction version: ${transaction.version}"); // Version: 2
+  print("Transaction Inputs: ${transaction.inputs.length}"); // 2 Inputs in this transaction
+  print("Transaction Outputs: ${transaction.outputs.length}"); // 2 Outputs in this transaction
 
-    test('base58 test', () {
-      expect(Base58Encode([0]), '1', reason: 'In Base58 the digit 0 is suppressed not to be confused with the capital O. Hence the value 0 is represented by the digit 1.');
-      expect(Base58Encode([57]), 'z');
-      expect(Base58Decode('3yQ').toUint8List.buffer.asByteData().getUint16(0), 9999);
-    });
+  final TransactionInput input1 = transaction.inputs[0];
 
-  });
+  print("Input1 TX ID reference: ${input1.refTXID.toHex}"); // TXID this input is referring: 55a736179f5ee498660f33ca6f4ce017ed8ad4bd286c162400d215f3c5a876af
+  print("Input1 previous transaction output index: ${input1.outIndex}"); // Index of the output associated to this input from the referred TXID: index = 0
+  print("Input1 scriptSig: ${input1.scriptSig.toHex}");
+  print("Input1 nSequence: ${input1.nSequence.toHex}");
 
+  final TransactionOutput output1 = transaction.outputs[0];
+
+  print("Output1 bitcoin value in satoshi: ${output1.nValue}"); // Amount of bitcoin in satoshi: 320000
+  print("Output1 scriptPubKey: ${output1.scriptPubKey.toHex}"); // scriptPubKey: 76a914e993470936b573678dc3b997e56db2f9983cb0b488ac
 }