+++ lib/src/model/bitcoin/transaction_input.dart
@@ -1,9 +1,21 @@
-class TransactionInput {
-  // Previous transaction ID (hash256 of prev trx: 4 Bytes little endian)
-  final int refId;
-  // • Previous transaction index
-  // • ScriptSig
-  // • Sequence
+import 'dart:typed_data';
 
-  TransactionInput(this.refId);
+import 'package:btc_sdk/btc_sdk.dart';
+
+class TransactionInput {
+  // Previous transaction ID (hash256 of prev trx: 32 Bytes little endian)
+  final Uint8List refTXID; // Uint8List(32);
+  // Index of the output to be unlocked
+  final int outIndex; // uint32
+  final Uint8List scriptSig;
+  final int nSequence; //uint32
+
+  TransactionInput({required this.refTXID, required this.outIndex, required this.scriptSig, this.nSequence = 0xFFFFFFFF});
+
+  Uint8List get toUint8List =>
+      refTXID
+          .concat(outIndex.to32Bits())
+          .concat(VarInt.fromValue(scriptSig.length).toUint8List)
+          .concat(scriptSig)
+          .concat(nSequence.to32Bits());
 }