+++ lib/src/model/bitcoin/transaction_output.dart
@@ -1,0 +1,9 @@
+import 'dart:typed_data';
+
+class TransactionOutput {
+  // Uint64 (8 Bytes) bitcoin amount in satoshi
+  final int nValue;
+  final Uint8List scriptPubKey;
+
+  TransactionOutput({required this.nValue, required this.scriptPubKey});
+}