+++ CHANGELOG.md
@@ -1,3 +1,16 @@
 ## 0.1.0
 
 - Initial version.
+
+## 0.1.1
+
+- Updating documentation
+
+## 0.1.2
+
+- Updating documentation
+
+## 0.1.3
+
+- improvement of the HDWallet interface: 
+  - create HDWallet directly using a mnemonic.